webapp/requirements.txt
django==4.1.3
djangorestframework==3.14.0
drf-spectacular==0.24.2
requests==2.28.1
validators==0.20.0
django-environ
uwsgi
setuptools==57.0.0
pip==21.1.2
asgiref==3.5.2